Move DHCP foldet to UDP
