+++ Aura Operating System/Aura_OS/System/Network/IPV4/IPPacket.cs
@@ -58,7 +58,7 @@ namespace Aura_OS.System.Network.IPV4
             }
             else if (NetworkStack.MACMap.ContainsKey(ip_packet.DestinationMAC.Hash))
             {
-                DHCP.DHCPPacket.DHCPHandler(packetData);
+                Network.UDP.DHCP.DHCPPacket.DHCPHandler(packetData);
             }
         }
