+++ Aura Operating System/Aura_OS/System/Network/IPV4/UDP/DHCP/DHCPPacket.cs
@@ -10,7 +10,7 @@ using System.Text;
 * PROGRAMMERS:      Alexy DA CRUZ <dacruzalexy@gmail.com>
 */
 
-namespace Aura_OS.System.Network.DHCP
+namespace Aura_OS.System.Network.UDP.DHCP
 {
     public class DHCPPacket : IPPacket
     {
