+++ Aura Operating System/Aura_OS/System/Network/IPV4/UDP/DHCP/DHCPRelease.cs
@@ -8,7 +8,7 @@ using Aura_OS.HAL;
 * PROGRAMMERS:      Alexy DA CRUZ <dacruzalexy@gmail.com>
 */
 
-namespace Aura_OS.System.Network.DHCP
+namespace Aura_OS.System.Network.UDP.DHCP
 {
     public class DHCPRelease : DHCPPacket
     {
