+++ Aura Operating System/Aura_OS/System/Network/IPV4/UDP/UDPPacket.cs
@@ -29,7 +29,7 @@ namespace Aura_OS.System.Network.IPV4.UDP
             {
                 if (udp_packet.SourcePort == 67) 
                 {
-                    Network.DHCP.DHCPPacket.DHCPHandler(packetData);
+                    Network.UDP.DHCP.DHCPPacket.DHCPHandler(packetData);
                     return;
                 }
                 else if (udp_packet.SourcePort == 53)
